fix bonus append issue
