+++ pipex/out
@@ -1,51 +1,2 @@
-paths[0] -> /opt/homebrew/opt/libpq/bin
-paths[1] -> /opt/homebrew/opt/ruby/bin
-paths[2] -> /Users/herbie/Library/pnpm
-paths[3] -> /Users/herbie/.yarn/bin
-paths[4] -> /Users/herbie/.config/yarn/global/node_modules/.bin
-paths[5] -> /opt/homebrew/opt/ruby/bin
-paths[6] -> /Users/herbie/.bun/bin
-paths[7] -> /opt/homebrew/opt/libpq/bin
-paths[8] -> /opt/homebrew/opt/ruby/bin
-paths[9] -> /Users/herbie/Library/pnpm
-paths[10] -> /Users/herbie/.yarn/bin
-paths[11] -> /Users/herbie/.config/yarn/global/node_modules/.bin
-paths[12] -> /opt/homebrew/opt/ruby/bin
-paths[13] -> /Users/herbie/.bun/bin
-paths[14] -> /opt/homebrew/opt/libpq/bin
-paths[15] -> /opt/homebrew/opt/ruby/bin
-paths[16] -> /Users/herbie/Library/pnpm
-paths[17] -> /Users/herbie/.yarn/bin
-paths[18] -> /Users/herbie/.config/yarn/global/node_modules/.bin
-paths[19] -> /opt/homebrew/opt/ruby/bin
-paths[20] -> /Users/herbie/.bun/bin
-paths[21] -> /opt/homebrew/opt/libpq/bin
-paths[22] -> /opt/homebrew/opt/ruby/bin
-paths[23] -> /Users/herbie/Library/pnpm
-paths[24] -> /Users/herbie/.yarn/bin
-paths[25] -> /Users/herbie/.config/yarn/global/node_modules/.bin
-paths[26] -> /opt/homebrew/opt/ruby/bin
-paths[27] -> /Users/herbie/.bun/bin
-paths[28] -> /Users/herbie/.nvm/versions/node/v16.15.1/bin
-paths[29] -> /opt/local/bin
-paths[30] -> /opt/local/sbin
-paths[31] -> /Library/Frameworks/Python.framework/Versions/3.10/bin
-paths[32] -> /Users/herbie/.local/share/solana/install/active_release/bin
-paths[33] -> /opt/homebrew/bin
-paths[34] -> /opt/homebrew/sbin
-paths[35] -> /usr/local/bin
-paths[36] -> /System/Cryptexes/App/usr/bin
-paths[37] -> /usr/bin
-paths[38] -> /bin
-paths[39] -> /usr/sbin
-paths[40] -> /sbin
-paths[41] -> /usr/local/go/bin
-paths[42] -> /Library/Apple/usr/bin
-paths[43] -> /Applications/Wireshark.app/Contents/MacOS
-paths[44] -> /Users/herbie/.cargo/bin
-paths[45] -> /Users/herbie/.fig/bin
-paths[46] -> /Users/herbie/Library/Application Support/JetBrains/Toolbox/scripts
-paths[47] -> /Users/herbie/go/bin
-paths[48] -> /Users/herbie/go/bin
-paths[49] -> /Users/herbie/go/bin
-paths[50] -> /Users/herbie/go/bin
+     63     170    1554
+      2       2      15
