+++ pipex/parse.c
@@ -6,7 +6,7 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/17 20:18:00 by herbie            #+#    #+#             */
-/*   Updated: 2023/03/27 19:26:26 by codespace        ###   ########.fr       */
+/*   Updated: 2023/03/29 12:36:20 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,10 @@
 #include "str.h"
 #include "structs.h"
 #include "io.h"
+#include "split.h"
 #include "error.h"
 #include <stdio.h>
+#include <unistd.h>
 
 /**
  * @brief The parse_args function takes in a pipex struct and the arguments
@@ -41,4 +43,36 @@ t_bool parse_args(t_pipex *pipex, int argc, char **argv)
 	if (ft_get_outfile(pipex, argv, argc) == -1)
 		return (false);
 	return (true);
+}
+
+/**
+ * @brief The ft_get_infile function takes in a pipex struct and the arguments
+ * passed to the program. It then parses the arguments and returns the full
+ * program path.
+ *
+ * @param cmd
+ * @param envp
+ * @return char*
+ */
+char *ft_find_path(char *cmd, char **envp)
+{
+	int i;
+	char **paths;
+
+	i = -1;
+	while (envp[++i])
+		if (ft_strncmp("PATH=", envp[i], 5) == 0)
+			paths = ft_split(envp[i] + 5, ':');
+	i = -1;
+	while (paths[++i])
+	{
+		char *path = ft_strjoin(paths[i], "/");
+		char *full_path = ft_strjoin(path, cmd);
+		free(path);
+		if (access(full_path, F_OK) == 0)
+			return (full_path);
+		free(full_path);
+	}
+	//! FIX FREE(paths)
+	return (NULL);
 }
